+++ src/cosmology/api/__init__.py
@@ -1,11 +1,19 @@
 """The Cosmology API standard."""
 
-# LOCAL
 from cosmology.api.background import BackgroundCosmologyAPI
 from cosmology.api.compat import (
     BackgroundCosmologyWrapperAPI,
     CosmologyWrapperAPI,
     StandardCosmologyWrapperAPI,
+)
+from cosmology.api.components import (
+    BaryonComponent,
+    DarkEnergyComponent,
+    DarkMatterComponent,
+    GlobalCurvatureComponent,
+    MatterComponent,
+    NeutrinoComponent,
+    PhotonComponent,
 )
 from cosmology.api.constants import CosmologyConstantsAPINamespace
 from cosmology.api.core import CosmologyAPI
@@ -16,6 +24,14 @@ __all__ = [
     "CosmologyAPI",
     "BackgroundCosmologyAPI",
     "StandardCosmologyAPI",
+    # components
+    "GlobalCurvatureComponent",
+    "MatterComponent",
+    "BaryonComponent",
+    "NeutrinoComponent",
+    "DarkEnergyComponent",
+    "DarkMatterComponent",
+    "PhotonComponent",
     # wrappers
     "CosmologyWrapperAPI",
     "BackgroundCosmologyWrapperAPI",